requirements.txt
alembic==1.15.2
amqp==5.3.1
bcrypt==4.0.1
billiard==4.2.1
blinker==1.9.0
cachelib==0.9.0
celery==5.3.4
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dnspython==2.7.0
email-validator==2.0.0.post2
Flask==2.3.3
Flask-Caching==2.1.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-Principal==0.4.0
Flask-Security-Too==5.3.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
gunicorn==21.2.0
idna==3.10
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.5.2
Mako==1.3.9
MarkupSafe==3.0.2
numpy==2.2.4
packaging==24.2
pandas==2.2.3
passlib==1.7.4
prompt_toolkit==3.0.50
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
redis==5.0.1
six==1.17.0
SQLAlchemy==2.0.21
typing_extensions==4.13.0
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
Werkzeug==2.3.7
WTForms==3.2.1
